rest: moving rest api fetcher to extensions
