+++ contrib/client_ssl_auth/filters/network/source/BUILD
@@ -27,7 +27,7 @@ envoy_cc_library(
         "//source/common/common:enum_to_int",
         "//source/common/http:headers_lib",
         "//source/common/http:message_lib",
-        "//source/common/http:rest_api_fetcher_lib",
+        "//source/extensions/config_subscription/rest:rest_api_fetcher_lib",
         "//source/common/http:utility_lib",
         "//source/common/json:json_loader_lib",
         "//source/common/network:cidr_range_lib",
