+++ source/extensions/config_subscription/rest/BUILD
@@ -1,6 +1,7 @@
 load(
     "//bazel:envoy_build_system.bzl",
     "envoy_cc_extension",
+    "envoy_cc_library",
     "envoy_extension_package",
 )
 
@@ -20,6 +21,7 @@ envoy_cc_extension(
         "//test:__subpackages__",
     ],
     deps = [
+        ":rest_api_fetcher_lib",
         "//envoy/config:subscription_interface",
         "//envoy/event:dispatcher_interface",
         "//source/common/buffer:buffer_lib",
@@ -30,9 +32,23 @@ envoy_cc_extension(
         "//source/common/config:type_to_endpoint_lib",
         "//source/common/config:utility_lib",
         "//source/common/http:headers_lib",
-        "//source/common/http:rest_api_fetcher_lib",
         "//source/common/protobuf",
         "//source/common/protobuf:utility_lib",
         "@envoy_api//envoy/service/discovery/v3:pkg_cc_proto",
+    ],
+)
+
+envoy_cc_library(
+    name = "rest_api_fetcher_lib",
+    srcs = ["rest_api_fetcher.cc"],
+    hdrs = ["rest_api_fetcher.h"],
+    deps = [
+        "//envoy/event:dispatcher_interface",
+        "//envoy/runtime:runtime_interface",
+        "//envoy/upstream:cluster_manager_interface",
+        "//source/common/common:enum_to_int",
+        "//source/common/config:utility_lib",
+        "//source/common/http:message_lib",
+        "//source/common/http:utility_lib",
     ],
 )
